operators/vgg16-encoder/requirements.txt
tensorflow==1.15.2
Keras==2.3.1
grpcio==1.27.2
pillow
grpcio-reflection
